更新1.2.0 支持.Net 5，FreeSql升级到2.0.101，所有Nuget包升级到最新 IdentityServer4升级到V4.1.1 删除不必要的视图文件，只保留登录和登出接口
