+++ Admin.IdentityServer/Admin.IdentityServer.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net5.0</TargetFramework>
     <Authors>xiaoxue</Authors>
     <Description>中台Admin统一认证授权服务中心</Description>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
@@ -29,25 +29,25 @@
   </ItemGroup>
   
   <ItemGroup>
-    <PackageReference Include="AspNetCore.HealthChecks.UI.Client" Version="3.1.1" />
-    <PackageReference Include="IdentityServer4" Version="3.1.3" />
+    <PackageReference Include="AspNetCore.HealthChecks.UI.Client" Version="3.1.2" />
+    <PackageReference Include="IdentityServer4" Version="4.1.1" />
     <PackageReference Include="Microsoft.AspNetCore.Diagnostics.HealthChecks" Version="2.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="3.1.6" />
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="5.0.0" />
     <!--<PackageReference Include="Microsoft.AspNetCore.Mvc.Razor.RuntimeCompilation" Version="3.1.6" />-->
-    <PackageReference Include="NLog" Version="4.7.2" />
-    <PackageReference Include="NLog.Web.AspNetCore" Version="4.9.2" />
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="5.5.1" />
+    <PackageReference Include="NLog" Version="4.7.5" />
+    <PackageReference Include="NLog.Web.AspNetCore" Version="4.9.3" />
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="5.6.3" />
     <!--<PackageReference Include="AutoMapper" Version="10.0.0" />
     <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="8.0.0" />-->
     <PackageReference Include="UAParser" Version="3.1.44" />
     
-    <PackageReference Include="FreeSql" Version="1.6.0" />
-    <PackageReference Include="FreeSql.Repository" Version="1.6.0" />
+    <PackageReference Include="FreeSql" Version="2.0.101" />
+    <PackageReference Include="FreeSql.Repository" Version="2.0.101" />
     <!--<PackageReference Include="FreeSql.Provider.MySql" Version="1.6.0" />-->
     <!--<PackageReference Include="FreeSql.Provider.PostgreSQL" Version="1.6.0" />-->
     <!--<PackageReference Include="FreeSql.Provider.SqlServer" Version="1.6.0" />-->
     <!--<PackageReference Include="FreeSql.Provider.Oracle" Version="1.6.0" />-->
-    <PackageReference Include="FreeSql.Provider.Sqlite" Version="1.6.0" />
+    <PackageReference Include="FreeSql.Provider.Sqlite" Version="2.0.101" />
   </ItemGroup>
   
   <ItemGroup>
