+++ Admin.IdentityServer/Config.cs
@@ -9,19 +9,27 @@ namespace Admin.IdentityServer
 {
     public static class Config
     {
-        public static IEnumerable<IdentityResource> Ids =>
+        public static IEnumerable<IdentityResource> IdentityResources =>
             new IdentityResource[]
             {
                 new IdentityResources.OpenId(),
                 new IdentityResources.Profile()
             };
 
-        public static IEnumerable<ApiResource> Apis =>
+        // v4新增
+        public static IEnumerable<ApiScope> ApiScopes =>
+            new ApiScope[] {
+                 new ApiScope("admin.server.api")
+            };
+
+        public static IEnumerable<ApiResource> ApiResources =>
             new ApiResource[]
             {
                 new ApiResource("admin.server.api", "admin后端api")
                 {
                     UserClaims =  { ClaimTypes.Name, JwtClaimTypes.Name },
+                    //v4新增
+                    Scopes = { "admin.server.api" },
                     ApiSecrets = new List<Secret>()
                     {
                         new Secret("secret".Sha256())
@@ -37,7 +45,7 @@ namespace Admin.IdentityServer
                     ClientId = "admin.server.api",
                     ClientName = "admin后端api认证",
                     AllowedGrantTypes = GrantTypes.Implicit,
-                    ClientSecrets = new []{ new Secret("secret".Sha256()) },
+                    //ClientSecrets = new []{ new Secret("secret".Sha256()) },
                     RequireConsent = false, //同意
                     AllowAccessTokensViaBrowser = true,
                     AccessTokenLifetime = 3600 * 2, //2小时 = 3600 * 2
@@ -63,7 +71,7 @@ namespace Admin.IdentityServer
                     ClientId = "admin.ui",
                     ClientName = "admin前端",
                     AllowedGrantTypes = GrantTypes.Implicit,
-                    ClientSecrets = new []{ new Secret(appSettings.AdminUI.Secret.Sha256()) },
+                    //ClientSecrets = new []{ new Secret(appSettings.AdminUI.Secret.Sha256()) },
                     RequireConsent = false, //同意
                     AllowAccessTokensViaBrowser = true,
                     AccessTokenLifetime = appSettings.AdminUI.AccessTokenLifetime, //2小时 = 3600 * 2
