--- Admin.IdentityServer/Controllers/IdentityServer4/Account/AccountOptions.cs
@@ -1,25 +1,0 @@
-﻿// Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
-
-
-using System;
-
-namespace Admin.IdentityServer
-{
-    public class AccountOptions
-    {
-        public static bool AllowLocalLogin = true;
-        public static bool AllowRememberLogin = true;
-        public static TimeSpan RememberMeLoginDuration = TimeSpan.FromDays(120);
-
-        public static bool ShowLogoutPrompt = false;
-        public static bool AutomaticRedirectAfterSignOut = true;
-
-        // specify the Windows authentication scheme being used
-        public static readonly string WindowsAuthenticationSchemeName = Microsoft.AspNetCore.Server.IISIntegration.IISDefaults.AuthenticationScheme;
-        // if user uses windows auth, should we load the groups from windows
-        public static bool IncludeWindowsGroups = false;
-
-        public static string InvalidCredentialsErrorMessage = "Invalid username or password";
-    }
-}
