--- Admin.IdentityServer/Controllers/IdentityServer4/Account/RedirectViewModel.cs
@@ -1,12 +1,0 @@
-// Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
-
-
-
-namespace Admin.IdentityServer
-{
-    public class RedirectViewModel
-    {
-        public string RedirectUrl { get; set; }
-    }
-}
