--- Admin.IdentityServer/Controllers/IdentityServer4/Consent/ConsentInputModel.cs
@@ -1,16 +1,0 @@
-﻿// Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
-
-
-using System.Collections.Generic;
-
-namespace Admin.IdentityServer
-{
-    public class ConsentInputModel
-    {
-        public string Button { get; set; }
-        public IEnumerable<string> ScopesConsented { get; set; }
-        public bool RememberConsent { get; set; }
-        public string ReturnUrl { get; set; }
-    }
-}
