--- Admin.IdentityServer/Controllers/IdentityServer4/Consent/ConsentViewModel.cs
@@ -1,19 +1,0 @@
-﻿// Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
-
-
-using System.Collections.Generic;
-
-namespace Admin.IdentityServer
-{
-    public class ConsentViewModel : ConsentInputModel
-    {
-        public string ClientName { get; set; }
-        public string ClientUrl { get; set; }
-        public string ClientLogoUrl { get; set; }
-        public bool AllowRememberConsent { get; set; }
-
-        public IEnumerable<ScopeViewModel> IdentityScopes { get; set; }
-        public IEnumerable<ScopeViewModel> ResourceScopes { get; set; }
-    }
-}
