--- Admin.IdentityServer/Controllers/IdentityServer4/Consent/ScopeViewModel.cs
@@ -1,16 +1,0 @@
-﻿// Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
-
-
-namespace Admin.IdentityServer
-{
-    public class ScopeViewModel
-    {
-        public string Name { get; set; }
-        public string DisplayName { get; set; }
-        public string Description { get; set; }
-        public bool Emphasize { get; set; }
-        public bool Required { get; set; }
-        public bool Checked { get; set; }
-    }
-}
