--- Admin.IdentityServer/Controllers/IdentityServer4/Extensions.cs
@@ -1,34 +1,0 @@
-using System.Threading.Tasks;
-using IdentityServer4.Stores;
-using Microsoft.AspNetCore.Mvc;
-
-namespace Admin.IdentityServer
-{
-    public static class Extensions
-    {
-        /// <summary>
-        /// Determines whether the client is configured to use PKCE.
-        /// </summary>
-        /// <param name="store">The store.</param>
-        /// <param name="client_id">The client identifier.</param>
-        /// <returns></returns>
-        public static async Task<bool> IsPkceClientAsync(this IClientStore store, string client_id)
-        {
-            if (!string.IsNullOrWhiteSpace(client_id))
-            {
-                var client = await store.FindEnabledClientByIdAsync(client_id);
-                return client?.RequirePkce == true;
-            }
-
-            return false;
-        }
-
-        public static IActionResult LoadingPage(this Controller controller, string viewName, string redirectUri)
-        {
-            controller.HttpContext.Response.StatusCode = 200;
-            controller.HttpContext.Response.Headers["Location"] = "";
-            
-            return controller.View(viewName, new RedirectViewModel { RedirectUrl = redirectUri });
-        }
-    }
-}
