--- Admin.IdentityServer/Views/Account/AccessDenied.cshtml
@@ -1,8 +1,0 @@
-﻿
-<div class="container">
-    <div class="page-header">
-        <h1>Access Denied</h1>
-    </div>
-
-    <p>You do not have access to that resource.</p>
-</div>
