+++ Admin.IdentityServer/Views/Account/Login.cshtml
@@ -1,65 +1,72 @@
-@model LoginViewModel
+﻿@model Admin.IdentityServer.Input.LoginInput
 
-@section styles
-{
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="utf-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">
+    <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no" />
+    <title>登录-中台admin</title>
     <link rel="stylesheet" href="~/css/is4.min.css?v=1.1.1" />
-}
-
-<div class="header">
-    <span>Admin后台管理系统</span>
-    <div class="list">
-        <a href="https://www.zhontai.net" style="margin-right: 32px" target="_blank">文档中心</a>
-        <a href="https://www.admin.zhontai.net" style="margin-right: 26px" target="_blank">首页</a>
-    </div>
-</div>
-
-<div class="login-body">
-    <div class="row">
-        <div class="col-sm-6 login-desc">
-            <p class="main-title">
-                登录服务
-            </p>
-            <p class="secondary-title">
-                统一认证授权服务中心
-            </p>
-        </div>
-        <div class="col-sm-6">
-            @if (Model.EnableLocalLogin)
-            {
-<form asp-route="Login">
-    <div class="my-alert my-alert--error is-light"></div>
-    <input id="returnUrl" type="hidden" asp-for="ReturnUrl" />
-    <div class="form-group">
-        <label asp-for="UserName"></label>
-        <input class="form-control" id="userName" placeholder="账号" asp-for="UserName" autofocus autocomplete="off" value="user">
-        <span class="text-danger field-validation-error" id="lblUserName">请输入账号</span>
-    </div>
-    <div class="form-group">
-        <label asp-for="Password"></label>
-        <input class="form-control" type="password" id="password" placeholder="密码" asp-for="Password" autocomplete="off" value="111111">
-        <span class="text-danger field-validation-error" id="lblPassword">请输入密码</span>
-    </div>
-    @if (Model.AllowRememberLogin)
-    {
-<div class="form-group login-remember">
-    <label for="rememberLogin">
-        <input id="rememberLogin" asp-for="RememberLogin">
-        <strong>记住我</strong>
-    </label>
-</div>}
-    <div class="form-group">
-        <button id="btnLogin" class="btn" name="button" value="login">登&nbsp;&nbsp;录</button>
-    </div>
-</form>            }
+</head>
+<body class="is4">
+    <div class="header">
+        <span>Admin后台管理系统</span>
+        <div class="list">
+            <a href="https://www.zhontai.net" style="margin-right: 32px" target="_blank">文档中心</a>
+            <a href="https://www.admin.zhontai.net" style="margin-right: 26px" target="_blank">首页</a>
         </div>
     </div>
-</div>
 
-<div class="login-footer">
-    © 2020 All Rights Reserved. zhontai
-</div>
+    <div class="login-body">
+        <div class="row">
+            <div class="col-sm-6 login-desc">
+                <p class="main-title">
+                    登录服务
+                </p>
+                <p class="secondary-title">
+                    统一认证授权服务中心
+                </p>
+            </div>
+            <div class="col-sm-6">
+                <form asp-route="Login">
+                    <div class="my-alert my-alert--error is-light"></div>
+                    <input id="returnUrl" type="hidden" asp-for="ReturnUrl" />
+                    <div class="form-group">
+                        <label>用户名：</label>
+                        <input class="form-control" id="userName" placeholder="账号" asp-for="UserName" autofocus autocomplete="off" value="user">
+                        <span class="text-danger field-validation-error" id="lblUserName">请输入账号</span>
+                    </div>
+                    <div class="form-group">
+                        <label>登录密码：</label>
+                        <input class="form-control" type="password" id="password" placeholder="密码" asp-for="Password" autocomplete="off" value="111111">
+                        <span class="text-danger field-validation-error" id="lblPassword">请输入密码</span>
+                    </div>
+                    <div class="form-group login-remember">
+                        <label for="rememberLogin">
+                            <input id="rememberLogin" asp-for="RememberLogin">
+                            <strong>记住我</strong>
+                        </label>
+                    </div>
+                    <div class="form-group">
+                        <button id="btnLogin" class="btn" name="button" value="login">登&nbsp;&nbsp;录</button>
+                    </div>
+                </form>
+            </div>
+        </div>
+    </div>
 
-@section scripts
-{
+    <div class="login-footer">
+        © 2020 All Rights Reserved. zhontai
+    </div>
+
+    <script src="~/lib/jquery/jquery.js"></script>
     <script src="~/js/is4.es5.min.js?v=1.1.1"></script>
-}
+</body>
+</html>
+
+
+
+
+
+
