--- Admin.IdentityServer/Views/Device/Success.cshtml
@@ -1,6 +1,0 @@
-<div class="page-header">
-    <h1>
-        Success
-        <small>You have successfully authorized the device</small>
-    </h1>
-</div>
