--- Admin.IdentityServer/Views/Shared/Error.cshtml
@@ -1,40 +1,0 @@
-@model ErrorViewModel
-
-@{
-    var error = Model?.Error?.Error;
-    var errorDescription = Model?.Error?.ErrorDescription;
-    var request_id = Model?.Error?.RequestId;
-}
-
-<div class="error-page">
-    <div class="page-header">
-        <h1>Error</h1>
-    </div>
-
-    <div class="row">
-        <div class="col-sm-6">
-            <div class="alert alert-danger">
-                Sorry, there was an error
-
-                @if (error != null)
-                {
-                    <strong>
-                        <em>
-                            : @error
-                        </em>
-                    </strong>
-
-                    if (errorDescription != null)
-                    {
-                        <div>@errorDescription</div>
-                    }
-                }
-            </div>
-
-            @if (request_id != null)
-            {
-                <div class="request-id">Request Id: @request_id</div>
-            }
-        </div>
-    </div>
-</div>
