--- Admin.IdentityServer/Views/Shared/Redirect.cshtml
@@ -1,7 +1,0 @@
-@model RedirectViewModel
-
-<h1>You are now being returned to the application.</h1>
-<p>Once complete, you may close this tab</p>
-
-<meta http-equiv="refresh" content="0;url=@Model.RedirectUrl" data-url="@Model.RedirectUrl">
-<script src="~/js/signin-redirect.js"></script>
