--- Admin.IdentityServer/Views/Shared/_Layout.cshtml
@@ -1,23 +1,0 @@
-﻿@using IdentityServer4.Extensions
-@{
-    string name = null;
-    if (!true.Equals(ViewData["signed-out"]))
-    {
-        name = Context.User?.GetDisplayName();
-    }
-}
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="utf-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">
-    <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no" />
-    <title>登录-中台admin</title>
-    @RenderSection("styles",required: false)
-</head>
-<body class="is4">
-    @RenderBody()
-    <script src="~/lib/jquery/jquery.js"></script>
-    @RenderSection("scripts", required: false)
-</body>
-</html>
