--- Admin.IdentityServer/Views/Shared/_ScopeListItem.cshtml
@@ -1,34 +1,0 @@
-﻿@model ScopeViewModel
-
-<li class="list-group-item">
-    <label>
-        <input class="consent-scopecheck"
-               type="checkbox"
-               name="ScopesConsented"
-               id="scopes_@Model.Name"
-               value="@Model.Name"
-               checked="@Model.Checked"
-               disabled="@Model.Required" />
-        @if (Model.Required)
-        {
-            <input type="hidden"
-                   name="ScopesConsented"
-                   value="@Model.Name" />
-        }
-        <strong>@Model.DisplayName</strong>
-        @if (Model.Emphasize)
-        {
-            <span class="glyphicon glyphicon-exclamation-sign"></span>
-        }
-    </label>
-    @if (Model.Required)
-    {
-        <span><em>(required)</em></span>
-    }
-    @if (Model.Description != null)
-    {
-        <div class="consent-description">
-            <label for="scopes_@Model.Name">@Model.Description</label>
-        </div>
-    }
-</li>
