--- Admin.IdentityServer/Views/Shared/_ValidationSummary.cshtml
@@ -1,7 +1,0 @@
-﻿@if (ViewContext.ModelState.IsValid == false)
-{
-    <div class="alert alert-danger">
-        <strong>Error</strong>
-        <div asp-validation-summary="All" class="danger"></div>
-    </div>
-}
