+++ Admin.IdentityServer/appsettings.json
@@ -44,12 +44,12 @@
     //监听Curd操作
     "curd": false,
     //数据库类型 MySql = 0, SqlServer = 1, PostgreSQL = 2, Oracle = 3, Sqlite = 4, OdbcOracle = 5, OdbcSqlServer = 6, OdbcMySql = 7, OdbcPostgreSQL = 8, Odbc = 9, OdbcDameng = 10, MsAccess = 11
-    "type": 4,
+    "type": "Sqlite",
     //连接字符串
     //MySql "Server=localhost; Port=3306; Database=admindb; Uid=root; Pwd=pwd; Charset=utf8mb4;"
     //SqlServer "Data Source=.;Integrated Security=True;Initial Catalog=admindb;Pooling=true;Min Pool Size=1"
     //PostgreSQL "Host=localhost;Port=5432;Username=postgres;Password=; Database=admindb;Pooling=true;Minimum Pool Size=1"
     //Sqlite "Data Source=|DataDirectory|\\admindb.db; Pooling=true;Min Pool Size=1"
-    "connectionString": "Data Source=D:\\Admin.Core\\Admin.Core\\bin\\Debug\\netcoreapp3.1\\admindb.db; Pooling=true;Min Pool Size=1"
+    "connectionString": "Data Source=F:\\admin\\Admin.Core\\Admin.Core\\bin\\Debug\\net5.0\\admindb.db; Pooling=true;Min Pool Size=1"
   }
 }
