+++ Admin.IdentityServer/tempkey.jwk
@@ -1,0 +1,1 @@
+{"alg":"RS256","d":"i2slA-hKDhn5niL4Z6DgEiHgi3mfpHNcsZTc5IwlDr8FoqnnxbJSNc39UZuIfHABLXGgcnyiNo9ciZ5Co-lfSDDBmQ-1pcuc8q1IQktKOwprUH5Rnv0ONi4FisoOAGQC8ke6h0kK-_bMzqya6qvk_6-TqrMc3mWOhY8UBMx-spWrNNkAMHVg_pjnHsdfXtVDPi-uTtFr0Guth20YWhtCY1VNq73HYn5sRo_K6Wr1-uZ44cupG3zrmcsty_RZ9Zu0KeMGwzzh580UYF7q9viIjYhDub-xL5JKD7WBKUVgEeeEGvIKXaYjOspErrGWByKQ545SfQoEQATitzNol3mgtQ","dp":"Yh886Fw5eC2DU4jH4rq9F_AaWPn0WD_fqC_gTjooohIYF3Jj02nUShWJ9HfIBE0IeZr24iJcjRmC7AAZzfnCRBtdVqBtYjsF6uGpqiBMVda5E1s6DcV1yYymYclGUZJHO55gFY-VNuCKXYaPCHjxxKsO4ZSHFitkJlgs7C8ALR8","dq":"jksvZg_lhf_USEKe60pjrxlT_aU2MF9--TCtxi9s3113DtN_zqJ4rUY_dOeQLtzJjm4-cL7xjWLPO8T1CvelmdSpey5kx7MLCmQ3fVHKDTJifc6qnEeMB_fO2sQIL1X6gYxmq_4KgvVx0yR1IsG5gT9FVNIhHF70u_bkiTuV5W8","e":"AQAB","kid":"A841BBCFC0B6F7CC8F14D11C8E556736","kty":"RSA","n":"1-4L78tUZFTE3P402vCytPzHjpDpwU4nLCrZFvpEtShZTmKjvkNDBvbwJA_1yaOORXUR_KeodzdO9nO8R2Ny5wU1KLeqgy5HqAY2e4X9UNxA5XgYty6Pm2d-S7rRi7jAttGh95BbgRWrUqr3eKtg6QUh5ijh4wkrzT09fwIGDnHC9KyQ_BffKzUQZz_loDDFdZdIoSNblk5fcdRN2rE5UcqYQSI10d8lLRkrb1sx7yWWz48cogoBMv5CZxrhVAABQiBDn7ew_cedVU1puXGOQ585V_IulJ1GDFJuB_r5CeaWa6V7VjDVW0DnMqOu5iqH4dcND-E3WA2CiBD0xgHmzQ","p":"-hegQjL5KgjGu7ZmKgVQlMFeiyD_X_LbaX_pCtjFkfkSe9usTrWJJM5KMUy8cvFxKCrGrI0BI7ufPgGrsWH1w4sM-P-huQRzqyFjWW9SvwHpJEXdnymfMogE183Skam6ldrIBDfSRDdoaOf7SoVVeAdIiUSnVBBPqmvG5UmCVf8","q":"3QfU2r8XhgxtoXi5EyOkJRdyKBDd9syulRqsU4vusA-Sd_kM4eZtQKHQ_rCDrr98OvO34s-Wi9yFKf7NtHAPislt82o53JN13E48SQm14jXkE-7kS3hm8-4z2G6kqO-1GR6WnQwPOUWxEEHGY5ZgQ1TPnHBWAX44Q8kUJsLHOzM","qi":"dsoOgyWxMlXvZn_omMU2yoqcz04-dQqnnYNllngSA-XykKP0MNsC0kQsNaSClyocdE7O35vKwBzaaDoG00afcHz5q1PnDSoP_bbMsDVRYEm3U-lLBXmbJbWGXUzdP6MKxs2oiA42ryzF1mbt5yff_ym5ShgfAg6YlKfkLvuNhOk"}
